Remove array spread operator from webChat.js
The array spread operator used in the getQueuedApis function causes
Babel to create a helper function that in turn uses Array.from.
Array.from is not compatible with IE 11, which we want to support. Also,
removed the eslintrc rule which errors when the spread operator is not
used with bind.

Change-Id: I03e608128ba5af0fa99ddff1fb39719bbdba3883