Refactor attachment message rendering logic
To get all the right properties of the attachment element to render, we
were using a `renderConfig` object local to the main attachment render
function. We were also passing the values of `renderConfig` to other
helper functions. This would become a problem if we have to pass values
to a function after a function (skipping the usage of the values in the
first function).
Refactored the code to use a private property to hold the render config
which is available throughout the class.

Change-Id: I9f20bd90f66f201861a17fae09cf939d04e747cc