Add support for setting z-index of web chat launcher
Based on the z-index values of the client's web page, they might want to
update the value of z-index of web chat launcher and the chat widget. We
are going to add the option to set the z-index value under the
helpshiftConfig.widgetOptions object. The z-index passed with this gets
set to the launcher and (value + 10) gets set to the chat widget iframe.

Change-Id: Icb63510601407c0da974ca18b3eddfd0108c80d1
JIRA: https://helpshift.atlassian.net/browse/REQ-680