Set device id in the state and localstorage
When the app is initialized, along with the rest of the config data, the
device id needs to be set in the app state. If a device id is present in
the localstorage, it has to be used, otherwise a new one needs to be
created. The device id should never be cleared, e.g. on reset.

Change-Id: I323d2471e613e824e4a59979dfca0221c970166d