Set anon user id in state and localstorage
A default profile always exists in the system, which is denoted by the
anonymous user id. The anon user id is cleared only when the `clear
default profile` API is called.
If a userId is passed with the `helpshiftConfig` object, it is used to
communicate with the backend. Otherwise, the anon user id is used.

Change-Id: I2ac3f3236b7fb163414c0f744fe9d4a929fa6c4e