Set issue details in the state after config is processed
Once the config is processed by Web Chat, it needs to get issues from
the backend and set issue details (issue id) in the state. If an active
issue exists for the profile, polling for issue data and messages
starts. If an active issue doesn't exist, Web Chat waits for the widget
to maximize when a pre-issue gets created.

Change-Id: I87b0884bdcf4496b1e34826feeab0f62dcd1d2cd