Render reply box for input type plain text
As there are no validations on input type plain text, the user can enter
multiline text. If we keep input type text, the long text will clip the
user text once it goes beyond available width.
Note :- We are not directly rendering reply box component as it does not
supports label and error.

Change-Id: Id980cc2dcc49e23ff44be2acf45380941c7c1621