Default developerSetLanguage to `en-US`
With the unification release, all web chat instances will start
localizing strings based on the browser's language. There are cases when
a developer may not want to do this immediately.
After introducing the localization capabilities of web chat, we'll give
developers some time (~30 days) to use the `language` config option to
set developer's language. Until then if this value is not set, we'll
default the developerSetLanguage value to `en-US` - which is the
current language experience. If someone still wants to use it, they can
set `langauge` to `browserDefault` and then we will pass browser's
language to the backend.

Change-Id: Ic33129e70c5d37739381cc466ea5e35483ded113