Fix : Enable footer if preIssue is resolved
+ show start new conversation button when preIssue is resolved
+ rename handle preIssue footer method
+ change the name of flags

We have to explicitly enable footer when issue is resolved.
Reason being, for preIssue we wait only
for bot step (according to design).
So when issue is deflected i.e user accepts faq suggestion, the footer is
hidden and user will not be able to see 'Start new conversation' button.

Change-Id: I59e498c29b12e3e25c2f2ffc59e028212e917636