Change get-user-config XHR method to "POST"
The generated XHR payload can be large in size. This can be a problem
in case of GET request as the max size that browsers support varies
from 2KB to 8KB & ideal is 255 characters. POST can handle large
amount of payload as compared to the GET. This usually depends
on the server configuration.

Change-Id: Iefb3af5c6a957ea60f56b5ec1dbe5535d2ef437d