Add functionality to handle unregistered events
Description: When "user changed" event is fired the event handler
might or might not be present because of async behaviour of the post
messages. Because of this the "user changed" event happens before
actual registration of event and hence the handler is never called.

Fix: Create a register for events for which we want to maintain the
state of event even after it has been fired. The event will be
consumed by the handler when it is registered. If the event is never
registered then it stays in memory for as long as app is running.

patch update - 2

Change-Id: I2c09844cb801f8865dd5faab4ba88a6c0c78bdfc
JIRA: https://helpshift.atlassian.net/browse/FRON-3366