Stop using "*" as targetOrigin with postMessage
postMessage from web chat -> parent uses "*" for the first message it
passes to the parent (sdk_js_loaded). Subsquent calls use the parent
page origin stored in the state. We want to avoid using "*" even for the
first call. This is done by passing the parent origin as a query param
to the web chat iframe's URL and then reading it in the web chat app
before calling all postMessage requests from web chat -> parent.

Change-Id: Ifddde5108d11d5f49ede4dc071932a6e81c5a0b7