Fix behavior of "hide" API
If the "hide" API is called two times, and then "show" API is called
after that, the webchat widget doesn't show.
It's because when the "hide" API is called the second time,
we save the current webchat visibility in the state (which is "none"),
and we set the same state when the "show" API is called.
That's why the webchat doesn't show up.

This commit fixes this issue by not saving the webchat visibility state
if it is already hidden.

Fixes oncall https://helpshift.atlassian.net/browse/ONCALL-3710

Change-Id: I8a8b800e3f23ee9909fbe1c4d908a0bd5d8c6bf6