Add stylelint rules.
* Some rules disallow vendor prefixes. These rules have been added
because we use autoprefixer to add vendor prefixes, and manually added
vendor prefixes should be removed.

* Some rules disallow unknown values, selectors, pseudo elements etc.
These rules help in adding invalid CSS and prevent spelling mistakes.

* selector-pseudo-element-colon-notation enforces double colon notation
for pseudo elements.

* length-zero-no-unit enforces not adding unit for zero value.

* declaration-block-no-shorthand-property-overrides: Disallow shorthand
properties that override related longhand properties. E.g.

This is invalid
a { padding-left: 10px; padding: 20px; }

This is valid:
a { padding: 20px; padding-left: 10px; }

Some rules have been added as a warning. This is because the existing
SCSS violates these rules and needs to be fixed.

Change-Id: Ice3eddfc0969d98416d23e8ec18938c91dfb76c2