Dispatch an action to post message instead of calling it directly
The current postMessage module accesses store. When it is called during
unit tests (that uses a mock store), it tries to use the actual store
instead of the mock store. This breaks the tests. The correct way to
call the postMessage api (esp if it uses the redux store) is to do it
via dispatch.

This commit contains the following changes -
1. Create a new actions module (actions/postMessage) that mimics the
behavior of the exisitng extras/postMessage module. It uses
util/postMessage to call the postMessage api.
2. Delete extras/postMessage module because its responsibility is now
covered by the new module.
3. Remove the use of store from utils/postMessage. Pass the value of
parentPageOrigin (for which the store was being used) via the action.
4. Replace the usage of extras/postMessage with actions/postMessage in
all the modules. Like `postMessage (event)` -> `dispatch (postMessage
(event))`.

Change-Id: I1686811e77bb652be11df595a012adfe638aa307