Add an option to query-stringify xhr data
Update helpers/xhr's getPreparedXhrData to optionally query-stringify
the xhr data. This is in order to make xhr calls without using our xhr
untility. The xhr utility does this (query-stringify) internally. This
changes is needed for when we use another way to fire the xhr (e.g. via
fetch or axios).

This change also includes a way to pass custom headers to the
getCommonHeader method. This is needed again because today some headers
are added by the xhr util by default.

- Move skipPlatformId option to an object
- Add queryStringify option to getPreparedXhrData
- Add _queryStringify function (copied from the xhr util)

Change-Id: Ic1dd4c1d8c856eab6dad1473c3af91ac1e70c7d2