Remove unused dependency from helpers/analytics
Created this commit because we need to bump version and
release it to invalidate the browser cache.
Ops added a new header `Vary: origin` to not serve the requests
from the cache if it is from a different origin.
But existing requests would still be cached.
By releasing this commit, we would ensure that the browser
would not serve the requests from the cache because the
request URL would be different as we append the project
version to all the requests.

More info at following links:
https://helpshift.atlassian.net/browse/ONCALL-4231
https://helpshift.atlassian.net/browse/OPS-3383

Change-Id: Ie926d2ed305097dd273e23b7b15244ef3e9eff13