Fire newUnreadMessages event conditionally
Add the following conditions before firing this event -
- web chat iframe is not open

Lack of the this condition was causing the unread message event
to be fired even when widget was open. If a customer has their own
notification system, it becomes annoying for the end user to receive
notification even when they have seen the message.

Change-Id: I1da59d92966a527f2120237b3db650fe205059a2