Fire unread message event irrespective of widget state
We made a change recently to not fire unread message event if the widget
was in an open state. This may cause custom notification to not clear
the notification when the widget opens. Removing the condition to check
for widget state before firing the event.

Also, made a change to update the unread count UI only when the xhr to
mark messages as seen would be succeessful. Earlier, the UI state was
being update irrespective of the success or failure of this xhr.

Change-Id: I31a6c9dd6dfa9c9a2046183e6d793a0b9f7babfe