Add logging to app
This sends error logs to kibana if app or any of its child at any level
encounters a javascript error. This is done to prevent minimize errors
once React is upgraded to v16.

Stack trace is sent as query param of a GET request to a non existent
route. This route, just like any other route gets logged in kibana. We
can use the stack trace to debug the errors. Similar logging has been
implemented in moby with React 16 error boundaries.

Change-Id: Id9f61dc27b75b136d7bd4bc4c206515c4295dfba