Fix error - Active view not available on refresh
+ Reasons:
The active view is set in component did mount. But picker, as the child
uses accessibility functionality in its component did mount. The problem
is that the component did mount of the child gets called before parent
and parent is setting an active view. We are moving the set active view
method to the component will mount as this life cycle method is called
for parent component first and then the child
Change-Id: I5a1c63a7e0e1bfcaa2aab21b1cc693d92bf0b4d3