Rename name param in ax function to group
+ Group is used to indicate the selector category in meta list
Reason:
We are using highlevel group to combine multiple focusable items.
Earlier using "name" was confusing as the purpose of name was to
uniquely identify a single focusable item. But that is not the case now.
We need the information of a "group" to which a particular focusable
item will belong.

Change-Id: I0b30638344234d9320d8742ecc73739303bc2b04