Fix link focus in message list on new messages load
+ Problem -
Selectors created for links and messages are breaking when
DOM updates outside message list
The current getSelectorForElement function returns selectors that use
HTML, body & nested nth-child selectors. Which causes the selectors to
break when the order/index of any of the parent changes. An additional
attribute is defined in the function to set the target parent context.
When the target parent context is given, the function returns selectors
that refer only to the elements inside the given parent context. So,
the selectors don't break when the DOM changes outside the given parent
context.

Change-Id: I9437d3579cbc25777a49fbf77e8999ffee19f38a