Wait for end user's message for creating preissue
This is an important change for lazy preissue creation.
`reloadApp` of actions/common is the primary action that
- sets the app reset trigger
- resets the app, which in turn calls initializeConversation of
  appState. This function currently, creates a preissue, via another fn
  called `startNewConversation`.
Updated the `startNewConversation` fn by removing the create preissue
call and adding the enable reply box fn call. Also, updated the submit
handler of the reply box so that it creates a preissue if one doesn't
exist.
- Also updated some of the comments, wherever I felt it could be
  improved.

Change-Id: Id7cc4ad48cf3b473f675f157f9b25a5b733430b0