Use a session id with analytics event
Because with lazy preissue creation, a preissue id is not available when
firing analytics event. We are using a random string session id to
counter it. The value updates only when a new conversation is created.
Also removed the enqueue and flush logic for analytics event tracking
because a session id will be used now.

Change-Id: Ifd40800d3d87b0fcc2d0c6a2271fd865f02f674e