Add check for footer selector length when footer changed
+ Component did mount of a child is called before the parent.
Picker as a child adds its selectors in meta-list on componentDidMount.
Chat view footer as parent overrides the footer selectors to empty,
in case of the footer as the picker. To solve this problem,
we check if the footer selectors length is empty, don't update the
footer selector in meta list

Change-Id: Ib3c9471014f5cc3505652429dabbf04d82464187