Reset active index on back btn click in FAQ view
After visiting the faq, the active index is set to the back button
selector index. Next time, we visit another faq, it focused on the back
button. The expected behavior is to focus on FAQ wrapper. To fix this,
we are resetting the FAQ active index when view change from FAQ to CHAT

Change-Id: I0ab9a27c906630023b2a8d4a0049c71159f73da6