Bug Fix: WCAX-69- Tab/Shift-tab interaction on closed launcher btn
Tab/Shift tab on launcher tries to focus on the element in
webSdkFrame. If the element is not found, it tries to focus
on the next/prev element on tab/Shift-tab. But when the
webSdkFrame is not opened, the flat list in ax will be
empty. This results in infinite recursion and it gives
stack overflow error.

Change-Id: Ic3a95513b54fbbb64bc8a94d2c881fb0514d3ea6