Remove unused function in ax
+ When the message length count is changed, we are saving
currentlyfocused link selector and adding selectors in the ax
msg list. Later we are focusing on the saved selector to retain
the current focus. But, when new messages loaded, the
selectors value is changed as the msg selectors are generated
dynamically. Currently saved selector is not available while
focusing. So, these functions are not adding any value.

Change-Id: Ib493837e9af29d5b268282e07c350a57e51cd7a0