Fix flickering of outline on launcher button
Move pills ax functionality to the message list. In chat
view footer component & option pills, in there respective
component did mount and component did update, add selectors
in meta-list only when the respective component is rendered
& then focus an element. This will cover the case when
neither footer or pills is rendered. We are not focusing
anywhere in this case.

Change-Id: I0f56145117e100d530f5cbc513fb84dd2fb9fc8b