Conditionally remove launcher button group
While navigating using tab/shift-tab, focuses the
corresponding next/prev element. If the launcher button is
hidden from webchat config options, ax module tries to
focus on launcher but it's not present on DOM. Due to this
ax breaks and there is an error on console.
To fix this issue we are removing launcher button group from
all the views' meta-list when showLauncher is hidden.

Change-Id: I53caf7089e6de039db1efb27d5cf364aa929f800