Move ax init func to componentWillMount in viewWrapper
The component will mount of the view wrapper is called before component
did mount, which sets active view and generate the flatlist. When the
launcher button is hidden from webchat config options, the
componentDidMount calls ax init function which resets the metaList by
removing the launcher button group. Old flatlist tries to focus on
launcher but it's not present on DOM. Due to this ax breaks and there
is an error in console.
To fix this problem, we are moving ax init function in
ComponentWillMount above setActiveView.

Change-Id: I902b26cc0ae2bc9c0649965600ede5b091d11328