Fix missing focus when CHAT view page is reloaded
On click new conversation button, webchat is reloaded. The
componentDidMount of chatViewFooter sets the footer group only when the
footer is rendered. The componentDidUpdate only update selectors for
new footer or refreshed footer. For chat view metaList the footer group
is empty when footer is not rendered and component did mount is called.
To solve this problem, we check if the footer is not rendered
with prev props and rendered with new props, which states that footer
is rendered for the first time.

Change-Id: I0da2c50148ad9ad96dc28ad35ec651b2510e5cc4