Fix mobile devices not displaying full screen
Earlier we were adding messenger styles first and then setting widget
position. Due to this, the full screen styles were reset. Now we are
changing the sequence of both the functions.
Also we are reseting max width and max height properties in full screen
styles. This is required when updateHelpshiftConfig is called.

Change-Id: Ia9ff27e6d9efd16640745f795aa4671f8f3d42cf