Show error for invalid file types
We accept file type only if it matches the mime types and file types
received in the config call. If admin changes the file types and the
user hasn't reloaded the page, webchat will not be able to detect
file invalidity. In this case we will show error sent by BE.

Change-Id: Id591621b5ae1760ecf44988286c40a7418bb33fe