Handle new conversation button click
Earlier, clicking on the new conversation button would simply call the
create preissue API and the poller would start. Before this, we used to
reset the state because the state would be rehydrated with the config
call, then the create preissue call, and finally the poller.
This has changed now because we are not supposed to call the create
preissue API with the new conversation button click. Instead we are now
changing the footer to accept the user input and updating the app state
by resetting the issue specific values e.g. issueState, activeIssueId,
etc.

This commit also contains some of the refactoring according to the new
convention of writing actions i.e. using actions as events instead of
indicators to update one flag/value of the state.

- Removed action to set conversationStarted flag. Instead the new
  conversation started action (event) will update this flag.
- Removed startConversation action. This was being used from api.js and
  it would call startNewConversation if issueExists was false. Handing
  this case directly in api.js using startNewConversation.
- Converted startNewConversation to an action so that more actions could
  be dispatched from it. Earlier, it was using store.dispatch.

Change-Id: I3e277b219edb446fcc1a7146a74a3ac66bd65138