Add support for internal data in helpshift config
We are adding support to pass internal data to webchat via helpshift
config object.
We need to pass voice meta data in internal data to webchat in case of
deflection use case. This will be passed to create preIssue xhr so that
BE knows the webchat issue is created for deflection use case.

Change-Id: I3b74129221b118bc5a5d55b4293ac0c59bb712fc