Use keyDown event to determine keyboard interaction
We have been using the keyPress event to determine whether the keyboard
interaction is active or not. This led to inconsistent setting of the
flag for keyboard interaction activity -> leading to a bug where the Ax
border would appear around answer bot replies [1].
The bug is fixed by removing the global keypress event handler and
letting each view handle the click and keydown events. The reply box
component also handles the space character keydown correctly with this
change.

[1] https://helpshift.atlassian.net/browse/HSBUG-132

Change-Id: I4a2faedd8c41e39c2aac374d15f309f60389c6d4