Show error for invalid file types in OOBH
Show file invalid error for each attachment, if that attachment
has an error. Before this change, we were showing file invalid
error at the wrapper of the attachments which fail to indicate
which file has an error.
If mime type list changes in the backend after webchat is loaded,
webchat will not be able to detect file invalidity. In this case,
we will show error sent by BE received in failed post issue xhr.

Change-Id: I499220e733813819320ba49512beccbf8f526eaa