Fix eslint make task for large number of files
All the changed files are passed to eslint gulp task through "gulp" npm
script. When a large number of files are changed, all these files are
passed as arguments to the npm script which are passed to gulp task.
NPM script was crashing for such huge number of arguments, although
gulp task was able to handle it.

This fix moves logic to find changed files in npm script itself, so we
don't need to pass any argument to npm script.

+ Remove xargs because it is not required
+ Move logic to remove "resources/" from filepath into bash

Change-Id: Id25dfdb4f532d5cbcfcd60ad6aacf15585288470