Check if attachment is image by its content type
It is related to attachments in bots. Previously, to determine image
attachment we checked for attachment extension, we get extension from
the last period until the end. But the new URLs sent by backend have
metadata after extension, metadata can have periods that result to the
wrong extension. So to avoid such case, we check image validity by
content type.

Change-Id: Ifbd6401a15b279f3331ea848a8fbf31771de735e