Fire the SDK loaded event after the CSS is loaded
- Fire the SDK loaded event after the CSS is loaded.
The CSS decides the height of the few elements,
and the height of picker depends on that.
If the client uses the Helpshift("open") API, sometimes, the widget gets
mounted even before the CSS gets loaded, resulting in wrong
height of the components.
With this change, the APIs would be triggered only after the CSS
is loaded (previously, they were getting triggered when only
JS was loaded).
This fixes https://helpshift.atlassian.net/browse/HSBUG-202

Change-Id: Ib26b01f6427d6a9914b8f63e710742188abadb86