Fix FRON-3731
The unread count was incorrect because in some cases we get preIssue +
issue data in poller. We only used to mark messages as seen for
latest issue. So if any preIssue has unread messages those will never
be marked as read.
In this fix, we collects all the issue or preIssue which has unread
messages and marks them as seen.

Change-Id: If546adc2941388e3922d8ead7bff955dcad83ecc