Pass correct max height to intents widget
Reduce the text input's height from the DND wrapper's height,
and pass it to Intent's NestedPicker.
Because of the excess height, the user input was flickering
when the intents state changed from OPENED to CLOSED.
Also, because the reply box can take multi-line text, update the
intents widget height when the reply box height get changed.

Change-Id: Ic86e6239953842d08bbfe537ea3c0398a7f8d804