HSBUG-396: Fix issue of intents not visible on mobile safari
We need to pass max height to picker when its in opened state for iOS
safari because safari pushes the entire webpage when keyboard is open.
This causes the smart intents to hide above the screen and user is not
able to see/select the intents. Restricting height in safari ensures
even after opening keyboard the intents are displayed to end user.

More info: https://helpshift.atlassian.net/browse/HSBUG-396

Change-Id: Ib7433906a12f20c6ecd35cb365bf0807da73a7e5