Send refers key with send message request only for bot messages
Before this change, web chat used to send the `refers` key with all send
message API calls. Backend uses this value only for bot messages.

The refers key was resulting in bad requests (400) when a message was
sent on a closed issue. The bad request was mainly due to the value of
the refers key. As a solution, we are removing this key for non-bot
messages because it's unnecessary anyway.

Change-Id: I3541974440942f8774ff360ca3529cb8f6bd5248
JIRA: https://helpshift.atlassian.net/browse/HSBUG-17