Handle 504 time out error for preissues API call
Currently, we have a generic retry handler for preissues API failure. It
tries to start a new conversation. This flow is not applicable to 504
time-out errors.

In case of 504 time-out errors, backend still creates the preissue
entity. Starting the poller is correct way to get out of this case.

Change-Id: Id9d8aa420225cc00ceb801da8eb19fa2f64080c6
JIRA: https://helpshift.atlassian.net/browse/HSBUG-481