Update state for preissue flow via event based actions
We were updating the state via setters directly from the action aysnc
functions. We do this elsewhere in the app as well, but we are trying to
move to an event based action system, where the async action dispatches
an event like action and the reducers update the state accordingly. This
makes managing the state and debugging via the dev tool easier.

This commit updates the create preissue flow.
- Add a new action creators module for chat view. The current common
  action creators module was getting too big.
- Rename createIssue action creator to createPreissueSuccess. Update the
  action type accordingly.
- Remove a todo that sets a fallback issue type. This was added with the
  smart intents feature and is not applicable anymore.

Change-Id: Ic533e11b391292f731268bc9d40b99ee5cb9bf99