Clear errors from chat view state when a poller succeeds
We had an event based action which was dispatched on every poller
success. Currently, it was only clearing the error from the chat view
state. Ideally this action should be managing more values in the state.

Dispatching an action just to clear the errors is unnecessary. So for
now, we will dispatch an action specifically to clear error from the
chat view state, only if an error exists.

Change-Id: I4ab2c51b3150e0c1353625fc3e24b643f8cd78e8