Reset widgetShouldAutoOpen flag on widget open
Currently this flag is used only with re-enagement and it gets reset on
page load - with the get config flow. There was anyway a bug in the
reset flow (the flag was not being read from appState). Moved the reset
logic to the widget open flow because once the widget auto-opens we can
safely reset the flag. This is needed for the message with actions
feature where the widget should auto open for non-mobile devices if a
link is clicked (given the link is configured to open in the same tab).

Change-Id: I972821e7d7dc59833faea89507a9bcbcd7907675