Auto open web chat for action link click
When an action card link opens in the same tab AND if web chat is not
open in full screen mode, auto open the web chat widget on the new page
(if web chat exists on that page).

Also, moved the action to set the widgetShouldAutoOpen flag to
actionCreators.js from actions/appState because it's now used in
actions/chatView as well.

Change-Id: Ibaeb8d8ae486511ce3991fce4611bca3591a2b44