Fix height of intent picker in case iOS liteSDK
+ We need to pass max height to picker when its in opened
state for iOS liteSDK because iOS pushes the entire
webpage when keyboard is open. This causes the smart intents
to hide above the screen and user is not able to see/select
the intents. Restricting height in iOS ensures that even
after opening keyboard the intents are displayed to end user.

More info: https://helpshift.atlassian.net/browse/HSBUG-396

Change-Id: If2488eacee635d7532f271b5b007905a328fa087