Fire conversationStart event on create preissue success
Earlier, the conversationStart event was getting fired when a user reply
was successful for the empty message with text input type of message.
This type of message is no longer used because of the preissue
optimization. Moved the conversationStart event to the success callback
of the create preissue API call.

Change-Id: I8485ac8ee48e465feb5dc5146328b8535b72d11f
JIRA: https://helpshift.atlassian.net/browse/CSR-3826