Remove pushTokenSyncMap from appState
+ We don't need to store the mapping of user to its synced
push token as the liteSdk will send the push token only when
it needs to synced (ie. either for an ongoing issue or at
the time issue creation)
+ Earlier we kept the pushTokenSyncMap as a validation to
avoid any unnecessary syncs but now that the lite sdk will
be unregistering the token whenever the user logs out.
Hence, webchat can't keep the push token sync mapping as
after the unregisteration of token, the webchat would still
have the token registered for that user and would not
re-register it when lite-sdk sends the push token to webchat
(which would result in no push notifications for that user)

Change-Id: I5518a7b8534fa1ca8a3de65842d71ffc431f3811