Retain conversation when issue flag is stale in config
Problem -  When a new user logs into the webchat the issue_exists
flag is false. If issue is created by the user and then the
webchat is reloaded with localstorage, the webchat doesn't retain
the conversation, as issue_exists flag is false in the config
data present in the localstorage.
Solution - Maintain a flag in the localstoage which indicates
that the issue is created and the localstorage has stale
issueExists value and on refresh fetch config from backend
for the updated values.

Change-Id: Ic4eb8043cd06ec624fb000df3ba804a7f4bbc99f