Open action card links in a new tab for Lite SDK
When web chat is used in the Lite SDK context, it runs in the web view
within the mobile app. Opening action card links in the same tab (when
they are configured like so), blocks user from going back to the app
because of lack of navigation options. The solution is to open the
action card links in a new tab in Lite SDK irrespective of the
configuration.

Change-Id: I227488f21f673c840e214635a527a285a2fa0742