Log helpshiftConfig.uiConfig for feature analysis
With SDK X GA release (Apr 2021), we are removing
helpshiftConfig.uiConfig for UI customization. This capability is
moving to the dashboard. In order to seamlessly migrate the customers
that are using uiConfig, we want to analyze helpshiftConfig.uiConfig.

This XHR hits a non-existent endpoint with uiConfig as a param. We will
then check the nginx logs for analysis. This is the same approach that
we follow with React error boundary, check
gunpowder/utils/withErrorBoundary for details.

Change-Id: I3a6d12c6dedf2e396193b81845c6fdce2455d903